Fix issue with bold text and rename ad banners.
Fixed some issues in fonts.scss so that opensans regular and opensans bold
are now in the same font family.  Also, ghandi serif bold-italic is now in
the ghandi serif family. Using a "b" or a "strong" tag now causes fonts to
be bold and appropriately uses sans-serif or serif depending on context.

I also noticed that the ad banner's id was triggering my ad-block software.
I renamed all ID's that contained "ad" to contain "temp-banner" instead so
as not to trigger ad-block software.