Update README.md
Reduce the README to just the information necessary for people to make changes and submit PRs. The inner workings of the docs are in flux, so I removed all the explanations of that. I also noted that PRs must point at a version branch rather than master, as master is being deprecated/removed.