Correct links and remove in-page header links
Many links were left unspecified before launch so that we could correctly link them once we knew which form of link worked. Of those that were specified, only those with full URLs worked. The ones relying on the directory and S3 bucket structure to fill in the URL failed. All have been updated and corrected with full IURLs. Additionally, it was found that all links within a document to headers within that document (or to specific headers in  other pages) failed, as linking to headers below h2 is currently unsupported. We are working to change that, but until then, all links have been removed.