Seed static assets from the previous site
NOTE: Most of these files SHOULD NOT persist.
The motivation for taking the compiled/minified versions of the
coffee/js and scss/css rather than the core files is that I would
rather build the skeleton of this new site with the same style and
scripts first. Once we have enough of a site, we can start building
out the asset pipeline.
Even some of the (more) static assets, (e.g. the Versions and
DocsFontIcons webfonts, the un-compiled/minified version-bar.js and
.css files, etc.) should be revisited and modified before we call
this project "ready".