Prototype the _default single.html layout
This is the basic page layout that will be used to render any non-
specified content. At this point, it will correctly render the
top-banners, the side bar (thought the navigation list remains
empty), the basic Markdown content styled Basho-ish, and the
copyright footer.

Please be aware of the numerous TODO notes -- this IS NO a finished
product.

To accomplish this, a few images were added to static/images/shared,
and the all.css was slightly modified;
    * image urls pointing to /shared/latest/images/layout/ were
      changed to /images/shared/
    * /shared/latest/images/layout/logo.png specifically was changed
      to /images/shared/basho_docs_logo.png
    * That rule was expanded with
        background-size: contain;
        background-repeat: no-repeat;