First pass at getting a Ruby workflow up.
Add .sass-cache to git ignore so we never have to worry about it.
Setup a .ruby-version with the latest version available to rbenv.
Setup a Gemfile with a handful of gems that look handy. This set is
based off of the Gemfile from the previous repository, and on my own
guesses of what will be useful.