First pass of the side navigation menu.
There's definitely work to be done yet, but this should work as a
proof-of-concept, at the least. The root-level menus barely react to
being the selected menu item, the CSS is still a mess, and I have
yet to test what non-content items do in this context.

I deviated from the Middleman logic to allow items with sub-menus
(`<li><h4><span><a>` elements) to be links, in addition terminating
list elements (`<li><a>`). This necessitated some CSS changes to
allow those h4 elements to be (selectively) colored as a reaction to
being the selected item.

As a final note, much of this code makes me _very_ sad. Once Hugo
0.15 hit Brew, I'll be able to refactor a lot of the logic into a
much simpler recursive partial, which may necessitate some
additional CSS modifications, as the depth of the menu is currently
hard-coded to 3 levels. Hopefully the CSS will be extracted and
cleaned up by then, but it's an important note to keep in mind.