Remove versions.yml, version-bar.erb -- Replace with version-bar.coffee
We don't want to waste our time with the versions.yml from the
previous site and the version-bar.erb loaded that as part of its
compile step, so strip both of those, as well as references to
version-bar.js. Re-add the logic from version-bar.erb that we want
back into the site through version-bar.coffee, and compile that as
part of all.js.