Cop out and half-fix some JS
The community pages were breaking, because their URLs don't include a
version number. This only mattered because the bitcask capacity calc
was calling window.pageInVersion --- to check if the page was older
than 1.3.2.
As we're not deploying those pages anymore, I opted to remove the
check, rather than fix window.pageInVersion.