Added a note stating that Debian is only supported for development.
For each OS, added or updated the following in the installation sections:
- Link to the KV docs on configuring the limit on the number of open file handles
- Hard-coded the name of the installation package
- Corrected the installation command, where necesary
- Removed the statement on turning off AAE since it is disabled out of the box

Changed instances of riak_shell to riak-shell where appropriate and properly identified the configuration file as riak_shell.config.

Added a link to the instructions in the KV docs to perform when IP addresses are changed in the configuration files.

Capitalized keywords in SQL statements.