Fix a bug in, generally improve _fonts.scss
Switch to using placeholder classes (`%docs_icon`) rather than mixins or
normal classes to cut down on the amount of dead CSS generated, and to
communicate the intention behind the classes defined in this file.

Reorder and more clrearly specify the font-face declarations.

Fix a bug in the OpenSans family.