Add meta tags to <head> that describe the current pages' paths
This included some shuffling of the variable names in
common_variables.html. I'm not entirely happy with what I landed on, but
I figure a large change should be made at a later date to fully clean up
and homogenize those variablesi. (If nothing else, it will definitely
happen as part of the larger rework.)