Import inital single.html and _wireframe.scss
If these files looks a bit like they've sprung forth fully formed, it's
because they mostly did. This commit (and many that will follow) are
being pulled from a previous WIP branch as a migration/rebase effort.

NB. We're going to rework Hugo templates next to (rather than in-place
of) the current style. This will let me compare the old rendered page
to the new while running a single `rake hugo:server`.