Split out the content_navigation.html partial, and clean up around it
The content nav partial now takes a dict that is expected to include the
name of the menu to be rendered. This is to allow for e.g. "community"
to be manually specified.
A fair bit of whitespace cleanup also occurred, as well as cleaning up
which variabled were extracted from .Scratch.