Recreate the downloads.html layout
This change also renamed the os_meta.yaml to os_metadata.yaml, because
there really isn't any reason to be so terse.

Note that the styling isn't quite done. There should be a second pass as
a page-specific update after a first draft of the final style is in
place.

If this files looks a bit like it has sprung forth fully formed, it's
because it mostly did. This commit is being largely pulled from a
previous work as a migration/rebase effort. There were no significant
changes to the logic used to generate these downloads pages. The
formatting was cleaned up _significantly_, though.

SQUASH with downloads page