Rework the Version Selector, pretty much completely
The sm-down version hasn't changed much, but the md-up version is now an
overlay, rather than an inset element. There are also Edge Faders added
to the selector pane that will allow for horizontal scrolling, and more
Edge Faders added to the selector lists that will allow for vertical
scrolling.