Add note regarding `n_val`
There has been confusion regarding the `buckets.default.n_val` setting
available through `riak.conf`. It only changes the `default` bucket
type, not the default for new bucket types (naming is bloody hard, but
this was avoidable).

Add a note that the real default is 3 and cannot be changed, but that it
can be reconfigured each time a table is created. Separately I will file
a PR against the KV content to clarify the `buckets.default.n_val`
configuration setting.