Recalibrate colors, and switch to using semantic names
In _colors.scss, this included tweaking the descriptive color values to
more closely match the official branding colors and adding a bunch of
semantic names that made use of the descriptive names.
In (most of) the rest of the site, it included switching out descriptive
names for their semeantic counterparts.

I really hope this doesn't make the site more confusing.