Add clarifications around expected behavior in LIMIT/ORDER BY
Per Andrei, there is not a lot for someonet to do re: backend implementation of LIMIT/ORDER BY and their use of query buffers. However, there are a couple things they can pay attention to. Updated both 1.5.0 and 1.5.1 LIMIT/ORDER BY pages with that information and added ORDER BY to riakconf page.