Add note about manual URL requirement
With the site redesign, the way the redirects from the splash page work has changed and is limited to/dependent on the S3 bucket. Which means that, in order to get from the splash page to the KV/TS/CS index page, you must manually enter version number in the URL. 

Alternatively, you can sit around thinking it/your laptop is broken like I did. But I don't recommend this approach.